implement error models to invoice commands

+++ src/GasStation.Application/Commands/Invoice/Delete/DeleteInvoiceCommandHandler.cs
@@ -1,10 +1,12 @@
 using GasStation.Application.Common.Interfaces.Persistence;
 using MediatR;
+using ErrorOr;
+using GasStation.Domain.Errors;
 using Microsoft.EntityFrameworkCore;
 
 namespace GasStation.Application.Commands.Invoice.Delete;
 
-public class DeleteInvoiceCommandHandler : IRequestHandler<DeleteInvoiceRequest, DeleteInvoiceResponse>
+public class DeleteInvoiceCommandHandler : IRequestHandler<DeleteInvoiceRequest, ErrorOr<DeleteInvoiceResponse>>
 {
     private readonly IApplicationDbContext _dbContext;
 
@@ -13,16 +15,19 @@ public class DeleteInvoiceCommandHandler : IRequestHandler<DeleteInvoiceRequest,
         _dbContext = dbContext;
     }   
         
-    public async Task<DeleteInvoiceResponse> Handle(DeleteInvoiceRequest request, CancellationToken cancellationToken)
+    public async Task<ErrorOr<DeleteInvoiceResponse>> Handle(DeleteInvoiceRequest request, CancellationToken cancellationToken)
     {
         var invoice = await _dbContext.Invoices
-                      .FirstOrDefaultAsync(i => i.Title == request.Title, cancellationToken)
-                      ?? throw new Exception("The invoice with specified title doesn't exist!");
-        
+            .FirstOrDefaultAsync(i => i.Title == request.Title, cancellationToken);
+        if (invoice is null)
+        {
+            return Errors.Invoice.TitleNotFound;
+        }
+
         _dbContext.Invoices.Remove(invoice);
         var result = await _dbContext.SaveChangesAsync(cancellationToken);
         
         return result > 0 ? new DeleteInvoiceResponse() {IsDeleted = true} 
-            : throw new Exception("Something went wrong!");
+            : Errors.Database.Unexpected;
     }
 }
